requirements.txt
asgiref==3.2.10
Django==3.0.8
django-cors-headers==3.4.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
Pillow==7.2.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
